Move core stuff into Home Assistant integration (#22407)
* Move core stuff into Home Assistant integration

* Lint
