+++ tests/components/google_assistant/test_google_assistant.py
@@ -8,7 +8,7 @@ import pytest
 
 from homeassistant import core, const, setup
 from homeassistant.components import (
-    fan, cover, light, switch, lock, async_setup, media_player)
+    fan, cover, light, switch, lock, media_player)
 from homeassistant.components.climate import const as climate
 from homeassistant.const import CLOUD_NEVER_EXPOSED_ENTITIES
 from homeassistant.components import google_assistant as ga
@@ -56,7 +56,7 @@ def assistant_client(loop, hass, aiohttp_client):
 def hass_fixture(loop, hass):
     """Set up a Home Assistant instance for these tests."""
     # We need to do this to get access to homeassistant/turn_(on,off)
-    loop.run_until_complete(async_setup(hass, {core.DOMAIN: {}}))
+    loop.run_until_complete(setup.async_setup_component(hass, core.DOMAIN, {}))
 
     loop.run_until_complete(
         setup.async_setup_component(hass, light.DOMAIN, {
