+++ tests/components/homeassistant/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the Home Assistant integration to provide core functionality."""
